Implement Dungeon class, floors, stairs, items, inventory.
This is really two half-baked changed that happened to collide because I
was half-baked.

Actions the player can take are now part of Dungeon; floors still
control game logic, but act more like a simulator than an interface.

Stairs always lead to the second floor.  Erm.

The inventory is heavily busted, but it does work.  Items can be picked
up and are, probably, moved into the player's backpack.