Split up the DungeonFloor object some more.
Now it's a Map, which is explicitly built to be stupid and only care
about placement and geometry.  The main interface to a Map is a Tile,
but that's now generated kinda lazily and just examines the Map's
internals.

Map creation is done by a Fractor, keeping all that hard-coded test junk
out of the rest of the code.

Fixed a lot of API awkwardness here, I think, so huzzah.