Implement really bad AI.
I'm not quite happy with this yet, but I'm considerably more happy than
I have been, so I think it's worth checkpointing.

Actions are represented by classes, to keep their execution out of the
hands of particular actors.  This should, I think, simplify handling of
death and whatnot in the future, since the dying critter doesn't have to
go through contortions to ask the dungeon to delete itself.